Group_6_ass_2_2024
This is to comment on the participation that all team members had in the creation of this code. We all collaborate in making the code at https://colab.research.google.com/drive/16QmBlcm5AYOcgninV7SbRkmB8PqxQahg?usp=sharing#scrollTo=-8l9I_7TsBMl.
Likewise, the proactivity of all the members of the team made up of @elisangela-orihuela , @COBM0605 , @CamiDavi2001 and myself was crucial to accomplish the task. In the process, Camila had to retire by the time difference of the countries. She participated in the resolution of the exercises, but did not manage to upload her part with GitHub. Therefore, we had to add her part.